Don't need to use Lazy because we will always evaluate the expression
